pyproject.toml
[project]
name = "gss-member"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]
